Spostato le funzioni comuni in handlers/utils.php
